feat: support 3-pane diff mode and configs for window options
if you run `DiffDirs` command with three arguments, it starts
3-pane diff mode. The first two arguments are used to spawn left-top and
right-top panes respectively with nomodifiable. The last argument is
used to spawn a bottom pane with modifiable options.

This commit also include supporting configuration of window options. The
left_diff_opt is applied to the left pane and the right pane if it's in
3-pane diff mode. The right_diff_opt is applied to the right pane if
it's in 2-pane diff mode or the bottom pane if it's in 3-pane diff mode.